Move dataclasses to a separate module
